Rename notify_id to notify_template_id
