+++ CODE_GUIDE.md
@@ -37,18 +37,18 @@ def step_to_do_a_thing(context):
 
 Every context attribute used by the tests should be described here.
 
-| Attribute                 | Description                                                                                     |
-| ------------------------- | ----------------------------------------------------------------------------------------------- |
-| test_start_local_datetime | Stores the local time at the beginning of each scenario in an environment hook                  |
-| survey_id                 | Stores the ID of the survey generated and or used by the scenario                  |
-| collex_id                 | Stores the ID of the collection exercise generated and or used by the scenario                  |
-| emitted_cases             | Stores the caseUpdate DTO objects emitted on `CASE_UPDATE` events                          |
-| emitted_uacs              | Stores the UAC DTO objects from the emitted `UAC_UPDATE` events                                |
-| pack_code                 | Stores the pack code used for fulfilments or action rules                                        |
-| template                  | Stores the column template used for fulfilments or action rules                                  |
-| telephone_capture_request | Stores the UAC and QID returned by a telephone capture API call                                 |
-| notify_id                 | Stores the ID of the sms template used for the notify service                                 |
-| phone_number              | Stores the phone number needed to check the notify api                                 |
+| Attribute                 | Description                                                                     |
+| ------------------------- | ------------------------------------------------------------------------------- |
+| test_start_local_datetime | Stores the local time at the beginning of each scenario in an environment hook  |
+| survey_id                 | Stores the ID of the survey generated and or used by the scenario               |
+| collex_id                 | Stores the ID of the collection exercise generated and or used by the scenario  |
+| emitted_cases             | Stores the caseUpdate DTO objects emitted on `CASE_UPDATE` events               |
+| emitted_uacs              | Stores the UAC DTO objects from the emitted `UAC_UPDATE` events                 |
+| pack_code                 | Stores the pack code used for fulfilments or action rules                       |
+| template                  | Stores the column template used for fulfilments or action rules                 |
+| telephone_capture_request | Stores the UAC and QID returned by a telephone capture API call                 |
+| notify_template_id        | Stores the ID of the sms template used for the notify service                   |
+| phone_number              | Stores the phone number needed to check the notify api                          |
 
 ### Sharing Code Between Steps
